Create maybe-firejail.rs
maybe-firejail? What's that?

Background: firejailing of $PATH is currently done by creating symlinks
to firejail in $PATH before the real program. But this can cause
problems because firejail has some requirements which maybe to meet in a
sandbox (firejail-in-firejail has frequent trouble with e.g. private-etc
or nogroups).

That's where maybe-firejail can help. The symlink are created to
maybe-firejail instead of firejail. maybe-firejail then detects if it
should start the program with firejail (outside of a sandbox) or direct
(inside a firejail sandbox). Because maybe-firejail is that minimal, it
has far less requirements which can break it.

Unresoleved issue: How to inegrate a rust binary in a setuptools build
system? (binary not extension)

Compile (`-C target-cpu=native' is optional):
    rustc --edition=2018 -O -C target-cpu=native maybe_firejail.rs